finish norming and removed recursive files
