+++ makefile
@@ -6,17 +6,17 @@
 #    By: znichola <znichola@student.42lausanne.ch>  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/11/16 07:44:20 by znichola          #+#    #+#              #
-#    Updated: 2022/12/11 17:18:59 by znichola         ###   ########.fr        #
+#    Updated: 2022/12/11 19:31:41 by znichola         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME	=	push_swap
 
-FILES	=	main              stack_ops_0       stack_ops__1 \
-			process_inputdata stack_ops_1       stack_ops__2 \
-			stack_ops_2       util         \
+FILES	=	main                     stack_ops__1 \
+			process_inputdata        stack_ops__2 \
+			       util         \
 			 stack_ops__0      debug        \
-			init quicksort basic_sort util2
+			init quicksort basic_sort util2 process_inputdata2
 SRC		=	$(addsuffix .c, $(FILES))
 OBJ		=	$(SRC:.c=.o)
