+++ process_inputdata.c
@@ -6,36 +6,13 @@
 /*   By: znichola <znichola@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/05 20:20:43 by znichola          #+#    #+#             */
-/*   Updated: 2022/12/11 17:46:51 by znichola         ###   ########.fr       */
+/*   Updated: 2022/12/11 19:12:45 by znichola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-static int	countnums(char const *s)
-{
-	int		count;
-	int		flag;
-	int		old_flag;
-
-	flag = -1;
-	count = 0;
-	while (*s)
-	{
-		old_flag = flag;
-		if (*s == ' ')
-			flag = 0;
-		else
-			flag = 1;
-		if (old_flag != flag && flag == 1)
-			count++;
-		s++;
-	}
-	return (count);
-}
-
-// static int	ft_atoi_read(int *n, char **str)
-static int	ft_atoi_read(int *n, char **str)
+int	ft_atoi_read(int *n, char **str)
 {
 	int	s;
 
@@ -58,56 +35,12 @@ static int	ft_atoi_read(int *n, char **str)
 	return (SUCCESS);
 }
 
-static int	add_num(t_stack *s, int i, char **str)
-{
-	int	t;
-	int	e;
-
-	if (ft_atoi_read(&t, str))
-		return (FAILURE);
-	e = i;
-	while (++e < s->size) //TODO: swapped to ++e instead of e++
-		if (s->root_a[e] == t)
-			return (FAILURE);
-	s->root_a[i] = t;
-	return (SUCCESS);
-}
-
-int	find_solution(t_stack *s)
-{
-	int	flag;
-	int	i;
-	int	t;
-
-	s->r.solution = (int *)malloc(sizeof(int) * s->size);
-	if (!s->r.solution)
-		return (ERROR);
-	ft_memcpy(s->r.solution, s->root_a, s->size * sizeof(int));
-	flag = 1;
-	while (flag)
-	{
-		flag = 0;
-		i = -1;
-		while (i++ < s->size - 2) // overflow checked here!
-			if (s->r.solution[i] < s->r.solution[i + 1]) // overflow check
-			{
-				t = s->r.solution[i];
-				s->r.solution[i] = s->r.solution[i + 1];
-				s->r.solution[i + 1] = t;
-				flag = 1;
-			}
-	}
-	s->r.s_end = s->r.solution + s->size - 1;
-	return (SUCCESS);
-}
-
 int	init_ops(t_stack *s)
 {
 	s->o.root = (int *)malloc(sizeof(int) * s->size * 20);
 	if (!s->o.root)
 		return (ERROR);
 	s->o.c = s->o.root - 1;
-	// *s->o.c = 0; //TODO: this seems bad? and not needed and an overflow
 	return (SUCCESS);
 }
 
@@ -125,33 +58,7 @@ int	finish_setup(t_stack *s)
 	return (SUCCESS);
 }
 
-int	process_inputdata_old(t_stack *stack, char *str)
-{
-	int	i;
-	// int	t;
-
-	i = countnums(str);
-	stack->root_a = (int *)malloc(sizeof(int) * i);
-	if (!stack->root_a)
-		return (ERROR);
-	stack->root_b = (int *)malloc(sizeof(int) * i);
-	if (!stack->root_b)
-		return (freeret_1(ERROR, stack->root_a));
-	stack->size = i;
-	stack->a = stack->root_a + i - 1;
-	stack->b = stack->root_b - 1;
-	while (i--)
-	{
-		if (add_num(stack, i, &str))
-			return (FAILURE);
-		str++;
-	}
-	if (finish_setup(stack))
-		return (freeret_2(ERROR, stack->root_a, stack->root_b));
-	return (SUCCESS);
-}
-
-int	add_single_num(t_stack *s, int index, char *str)
+static int	add_single_num(t_stack *s, int index, char *str)
 {
 	int	number;
 
@@ -180,7 +87,6 @@ int	process_inputdata(t_stack *stack, char **str, int n)
 	while (n--)
 		if (add_single_num(stack, n, str[i++]))
 			return (freeret_2(ERROR, stack->root_a, stack->root_b));
-			// return (FAILURE); //TODO: what about free?
 	if (finish_setup(stack))
 		return (freeret_2(ERROR, stack->root_a, stack->root_b));
 	return (SUCCESS);
