+++ stack_ops__0.c
@@ -6,7 +6,7 @@
 /*   By: znichola <znichola@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/05 19:57:05 by znichola          #+#    #+#             */
-/*   Updated: 2022/12/08 18:54:02 by znichola         ###   ########.fr       */
+/*   Updated: 2022/12/11 19:31:00 by znichola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,27 @@
 
 /** only stack opperation */
 
+// return the height of stack x
+int	sh(t_stack *s, char x)
+{
+	if (x == 'a')
+		return (s->a - s->root_a + 1);
+	if (x == 'b')
+		return (s->b - s->root_b + 1);
+	return (ERROR);
+}
+
 // do x and y at the same time
 int	tg_(t_stack *s, int (*x)(t_stack *), int (*y)(t_stack *))
 {
 	if (x(s) + y(s) == FAILURE * 2)
 	{
-		// write(1, &"_warning: double fail .. impossible and dangerous\n", 51); //TODO: important things happen here
 		return (FAILURE);
 	}
 	return (SUCCESS);
 }
+		// write(1, &"_warning: double fail .. impossible and dangerous\n", 51);
+		//TODO: important things happen here
 // this funciton above is liable to be tricky
 // must be carefull to understand it's output
 // if there are 3 numbers or less it will always trigger the double fail.
@@ -49,8 +60,6 @@ int	sb_(t_stack *s)
 {
 	int	t;
 
-	// ft_printf("ksjhdf\n");
-	// print_stack(s);
 	if (!(sh(s, 'b') > 1))
 		return (FAILURE);
 	t = *(s->b);
