+++ stack_ops__1.c
@@ -6,13 +6,14 @@
 /*   By: znichola <znichola@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 10:58:14 by znichola          #+#    #+#             */
-/*   Updated: 2022/12/11 17:18:44 by znichola         ###   ########.fr       */
+/*   Updated: 2022/12/11 19:35:29 by znichola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-// pa (push a): Take the first element at the top of b and put it at the top of a.
+// pa (push a): Take the first element 
+// at the top of b and put it at the top of a.
 // Do nothing if b is empty.
 int	pa_(t_stack *s)
 {
@@ -24,7 +25,8 @@ int	pa_(t_stack *s)
 	return (SUCCESS);
 }
 
-// pb (push b): Take the first element at the top of a and put it at the top of b.
+// pb (push b): Take the first element
+// at the top of a and put it at the top of b.
 // Do nothing if a is empty.
 int	pb_(t_stack *s)
 {
@@ -42,12 +44,12 @@ int	ra_(t_stack *s)
 {
 	int	t;
 	int	h;
-	
+
 	h = sh(s, 'a');
-	if (h < 3) //TODO: changing this to not include 3!
+	if (h < 2)
 		return (FAILURE);
 	t = *s->a;
-	while (h-- > 1) //TODO: changed to be samller than 1!
+	while (h-- > 1)
 		s->root_a[h] = s->root_a[h - 1];
 	*s->root_a = t;
 	return (SUCCESS);
@@ -56,16 +58,15 @@ int	ra_(t_stack *s)
 // TODO: these two functions are horribly slow, 
 // I think, will see about optimisation later
 
-
 // rb (rotate b): Shift up all elements of stack b by 1.
 // The first element becomes the last one.
 int	rb_(t_stack *s)
 {
 	int	t;
 	int	h;
-	
+
 	h = sh(s, 'b');
-	if (h < 2) //TODO: changed to not include 3!
+	if (h < 2)
 		return (FAILURE);
 	t = *s->b;
 	while (h-- > 1)
